+++ src/config/config.ts
@@ -70,8 +70,8 @@ export namespace customConfig {
 		pushJobIntervals?: number
 		executeJobConcurrency?: number
 		keepRunning?: boolean
-		forceUpdate?: boolean
 		db: string
+		chunkSize: number
 		contractOwners: string[]
 	}